Add GOV.UK Apt repos in 00-base cloud-init snippet
Install the signing key and configure (one of) the GOV.UK Apt repos
(awscli).

This ensures we get an up-to-date awscli package during provisioning.
The awscli is required for instances to discover information about
themselves from the EC2 API that isn't available from the EC2 metadata
service, such as tags.